requirements.txt
uvicorn~=0.24.0
motor~=3.3.1
pydantic~=2.4.2
fastapi~=0.104.1
pymongo~=4.6.0
python-jose
pydantic_settings~=2.0.3
passlib==1.7.4
pytest
httpx